{{cookiecutter.dir_name}}/requirements.txt
Django==1.10.1
djangorestframework
markdown
django-filter
